requirements.txt
Django==1.11.7
dj-database-url==0.4.2
psycopg2==2.6.2
python-env==1.0.0
django-countries==5.0
pillow==4.3.0
djangorestframework==3.7.3
